+++ src/main/java/com/casseautomatiche/gdo/utilities/dto/ArticoloRequest.java
@@ -2,6 +2,7 @@ package com.casseautomatiche.gdo.utilities.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Data
@@ -12,6 +13,6 @@ public class ArticoloRequest {
     private Integer grammatura;
     private String unitaDiMisura;
     private UUID idReparto;
-    private Float prezzo;
+    private BigDecimal prezzo;
     private Integer stock;
 }
